add users and places controllers, setup places-controller
